[infra] Change upstream repo for abseil-cpp
Abseil (which is a DEP of dawn) has BUILD.bazel
files, but the ones in the chromium third_party
version are broken because there are some files
that are not included in the chromium clone
(fork?).

For example, internal/thread_annotations.h [1],
which is not found in [2].

We want to be able to use existing BUILD.bazel
rules as much as possible when building Skia
with Bazel, so this changes the repo we sync
abseil to be the one on GitHub (the source of
truth). The git revision I went with was the
most recent one and appeared to work locally.

I ran this idea by bajones@, who added the
DEP originally in [3], and got approval.

[1] https://github.com/abseil/abseil-cpp/blob/master/absl/base/internal/thread_annotations.h
[2] https://chromium.googlesource.com/chromium/src/third_party/abseil-cpp/+/refs/heads/main/absl/base/internal/
[3] https://skia-review.googlesource.com/c/skia/+/451638

Change-Id: I1a61907a86de08cbddc395bf261426237788c6a5
Bug: skia:12541
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/520696
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Eric Boren <borenet@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>