use fast case for blurred paths
If the paint has a blur mask filter, then we can adjust the sigma
of the filter to use the CTM adjustment case instead of
transforming the path. This allows reusing the mask cache for the
blured part of the drawing.

Tested on TOT Chromium with this CL patched in. Smooth as silk.

Bug: chromium:1304806

Change-Id: I5cda17f1769ef173e6734d22edbd3fef4ed8db6f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/521100
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Herb Derby <herb@google.com>