Simplify filtering in SkQPRunner.
Previously, we had a member variable `mShouldSkipTest` which existed to
track the filtered-or-not state of individual unit tests. Now, when a
unit test is filtered out, we just replace it with an empty Description
in the array of unit test descriptions.

`mShouldRunTestCount` was also unnecessary as it should always exactly
equal the number of child tests contained in the test-suite.

Bug: skia:13042
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/521636
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Leon Scroggins <scroggo@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>
Change-Id: I89841a3830edc8605fd64c0ea6ccb74528ffdc1e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/523181
Reviewed-by: John Stiles <johnstiles@google.com>