Fix EllipticalRRectEffect precision for large corner radii
Modifies rrects_* GMs to draw a clipped rrect with
a large radius. The _effect case would draw with a blurry
edge before the fixes.

1. Adds scaling to the elliptical effects key and activate
it when the rrect radii are large enough. This fixes the
precision issue with the elliptical effect variant.
2. Uses SkRRectPriv::IsSimpleCircular() to choose the
circular variant over elliptical, which uses a tolerance
between X and Y radii. This avoids the issue since the
circular effect doesn't suffer in this case.

Confirmed that both fixes work independently of each other.

Cq-Do-Not-Cancel-Tryjobs: true
Bug: skia:10456
Change-Id: Ifef63356e31f46ec1546684636d9667b90bcef9b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/521523
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>