Make anisotropic GM use mip maps
This GM was originally written against the old high-quality mode.
Given that it only draws at scale or downsamples, the modern equivalent
would be to use mip maps. However, when it was converted to
SkSamplingOptions it was made to use Mitchell.

The result is pretty ugly but hopefully this will help illustrate the
benefit of anisotropic filtering in comparison.

Bug: skia:13036
Change-Id: I2e5ef2d143fa148583391aace66b44fa59b7f202
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/522360
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Brian Salomon <bsalomon@google.com>