Use MTV Golo machines
This update the drivers and removes failing jobs (see [1] and [2]), but
does not replace them, as that would require backporting more changes
to add in the new Skolo GTX 360 machines.

[1] https://skia-review.googlesource.com/c/skia/+/525836
[2] https://skia-review.googlesource.com/c/skia/+/524001

There are some iPhone6 jobs removed here because they were removed
in m99, but that change was not cherry-picked back.

Bug: 1289453
Change-Id: Iea9d802cdfaffbf1f3f04fa0dd6ebcbe9a568e9d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/528079
Reviewed-by: Ravi Mistry <rmistry@google.com>