[graphite] Reuse local-to-device Transform between draws and clips
In profiling, calculating the inverse of the 4x4 every draw call or on
every clip operation was measurable, but it really only needs to be
done if the transform is stale before the next time it's used.

Code is a little cleaner now too, since we're staying skgpu:: types
more than before.

Bug: skia:12787
Change-Id: Ib6313f211c8ebd6a484878bf9975e5ad6a4e750f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/529126
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>