Generalize SkGlyph::fForceBW.
The field SkGlyph::fForceBW is currently used by the DirectWrite
SkScalerContext implementation to let generateImage know that
generateMetrics needed to fall back to aliased rendering in order to
load the glyph. This is part of a more general issue where
generateMetrics must generate the metrics based on some glyph
representation and then generateImage must use that same representation
to draw.

Currently this consistency comes from ensuring that generateMetrics and
generateImage make the same decisions, with the only real information
passed from generateMetrics to generateImage being the mask format.
However, there have always been multiple glyph represenations for a
given mask format (like outline masks vs. embedded bitmap masks) and now
there are four (soon to be five) different color formats. While the mask
format and bounds are enough to allocate the space for a glyph
rendering, this is not enough information for generateMetrics and
generateImage to agree on which glyph rendering to choose. The
additional required information has come from implicitly coding
generateImage to make the same decisions generateMetrics would have.
However, these decisions may have been based on information which is
expensive to compute, like probing the font for which representations a
glyph may have.

To transmit this extra information, formalize that generateMetrics may
state which decisions it made (which glyph representation it used) so
that generateImage can use this information explicitly instead of
implicitly. Do this by expanding fForceBW and renaming it
fScalerContextBits and documenting what it is used for.

Change-Id: I91ba747233447f4d7bea81b0dcde51e49fa705fe
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/529604
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>