Implement op cache for SPIR-V.
We have a new `writeInstruction` call which can cache instructions as
they are written. This allows us to reuse existing instructions instead
of recomputing the same things twice.

This feature is largely untapped in this CL. As a demonstration that it
works, we no longer check the `fNumberConstants` map when writing
numbers, and instead rely on the cache to deduplicate OpConstant
instructions. Followup CLs will leverage the feature more heavily.

Change-Id: I6c25d0c2db7888ace05c24f8f9961d25c91b93b5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/528817
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>