Fix PaintRotate and PaintSkew angle orientation
PaintRotate needs to rotate counter-clockwise, but was rotating
clockwise due to passing the font's value unchanged to
SkMatrix::RotateDeg. Negate the value to achieve counter-clockwise
rotation.

For PaintSkew, we compute the tangent from the values in the font and
set up a skew matrix with that. Since we're using the trigonometric
functions to set up the matrix, the angle works in counter-clockwise
direction.

So this previously worked correctly for skew in the x
direction, but not in the y-direction due Skia using a y-down coordinate
system. Negate the y-angle from PaintSkew before calculating the
tangent and setting up the matrix.

Bug: skia:13190
Cq-Include-Trybots: luci.skia.skia.primary:Test-Android-Clang-GalaxyS20-GPU-MaliG77-arm64-Release-All-Android_NativeFonts
Change-Id: I98329f9746bcc6e875c4f87982782470e0f0a17d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/529076
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>