Reorganize COLR font support with FreeType.
When using a COLR glyph representation there is no need to load the base
glyph to draw. If COLR glyph representation was chosen for metrics, the
image and drawable generation should use the same format. Separate the
drawing of COLR glyphs with Skia and drawing other glyphs with FreeType
data.

This is not intended to change behavior. However, this does incidentally
fix an issue where previously a COLRv0 glyph would use the metrics of
its last component instead of the metrics of the base glyph. The
specification requires that all COLRv0 layers have the same advance
width and the test fonts do this. However, other code like the FreeType
autohinter may kick in and cause the value to be slightly different.
Since COLRv1 requires that the base glyph advance be used, this change
standardizes on using the base glyph.

[0] https://docs.microsoft.com/en-us/typography/opentype/spec/colr#metrics-for-color-glyphs-using-version-0-formats

Change-Id: I0258c82d83076327cfb2c68954b9bb2683599e9c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/529804
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>