Change SkGlyph::fScalerContextBits to uint16_t.
This used to be uint8_t and was changed to uint32_t to soak up some
unused bits. However, it turns out that it was actually sitting in a
three byte region, so declaring this uint32_t accidentally made SkGlyph
larger. Changing this to uint16_t should allow it to soak up two of the
three unused bytes.

Bug: chromium:1315972
Change-Id: If02dc5d9f38db2715fb08a762246d16055e06214
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/529812
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>