Check FT_Face's glyph is outline before using it.
The current code assumes that the glyph loaded into the FT_Face has
outline format. However, nothing is actually checking this, it is
possible to get here which an SVG or bitmap backed glyph. Check that
glyph actually has an outline before trying to convert to SkPath.

Change-Id: I24f9f06a4ba6a3323a310a559a1548ff76cfdffb
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/530051
Reviewed-by: Herb Derby <herb@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>